Zabrana dodavanja premium knjige
Premestena u addToCart metod
closes #180

+++ E-library/Services/ShoppingCart/ShoppingCart.API/Controllers/CartController.cs
@@ -39,10 +39,10 @@ namespace ShoppingCart.API.Controllers
         [ProducesResponseType(typeof(Cart), StatusCodes.Status200OK)]
         public async Task<ActionResult<Cart>> GetCart(string username)
         {
-            //if (User.FindFirst(ClaimTypes.Name).Value != username)
-            //{
-            //    return Forbid();
-            //}
+            if (User.FindFirst(ClaimTypes.Name).Value != username)
+            {
+                return Forbid();
+            }
 
             var cart = await _repository.GetCart(username);
             return Ok(cart ?? new Cart(username));
@@ -53,10 +53,10 @@ namespace ShoppingCart.API.Controllers
         [ProducesResponseType(typeof(Cart), StatusCodes.Status200OK)]
         public async Task<ActionResult<Cart>> UpdateCart([FromBody] Cart cart)
         {
-            //if (User.FindFirst(ClaimTypes.Name).Value != cart.Username)
-            //{
-            //    return Forbid();
-            //}
+            if (User.FindFirst(ClaimTypes.Name).Value != cart.Username)
+            {
+                return Forbid();
+            }
 
             return Ok(await _repository.UpdateCart(cart));
         }
@@ -66,10 +66,10 @@ namespace ShoppingCart.API.Controllers
         [ProducesResponseType(typeof(void), StatusCodes.Status200OK)]
         public async Task<IActionResult> DeleteCart(string username)
         {
-            //if (User.FindFirst(ClaimTypes.Name).Value != username)
-           // {
-           //     return Forbid();
-            //}
+            if (User.FindFirst(ClaimTypes.Name).Value != username)
+            {
+                return Forbid();
+            }
 
             await _repository.DeleteCart(username);
             return Ok();
@@ -81,12 +81,14 @@ namespace ShoppingCart.API.Controllers
         [ProducesResponseType(typeof(void), StatusCodes.Status400BadRequest)]
         public async Task<IActionResult> Checkout(string username)
         {
-            //if (User.FindFirst(ClaimTypes.Name).Value != cartCheckout.Username)
-            //{
-            //    return Forbid();
-            //}
-
+            
             CartCheckout cartCheckout = new CartCheckout();
+
+            if (User.FindFirst(ClaimTypes.Name).Value != cartCheckout.Username)
+            {
+                return Forbid();
+            }
+
             List<Entities.CartItem> checkoutItems = new List<Entities.CartItem>();
             var cart = await _repository.GetCart(username);
             if (cart == null)
@@ -133,15 +135,15 @@ namespace ShoppingCart.API.Controllers
             {
                 return Forbid();
             }
-/*
-            if (book.IsPremium)
-            
-                if (!User.IsInRole("PremiumMember"))
-                
+
+            var book = result.Find(b =>  b.BookId == bookId);
+ 
+            if (book.IsPremium) { 
+                if (!User.IsInRole("PremiumMember")) { 
                     return Forbid();
                 }
             }
-*/
+          
             return Ok(result);
         }
